requirements.txt
asgiref==3.2.7
Django==3.0.5
django-crispy-forms==1.9.0
django-phone-field==1.8.0
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
djangorestframework~=3.11.0